Change return type of penguin_bravo to short
This conforms with the second exercise and will be used to show
CI pipeline failing. I also changed the interfaces.h to prevent
the CI from failing.

+++ penguin/interfaces.h
@@ -1,5 +1,5 @@
 #pragma once
 
 extern int penguin_alpha(void);
-extern int penguin_bravo(void);
+extern short penguin_bravo(void);
 extern int penguin_charlie(void);
